Fix app generation on windows to close #36
