+++ tools/generate/handlers/handlers.go.template
@@ -7,7 +7,7 @@ import (
 
 :	<@range $i, $v := .ctx.parents>\
 :	:	<@if $v.Import>\
-:	:		<@$v.Package> "<@join $.ctx.outputImport $v.Import>"
+:	:		<@$v.Package> "<@joinImp $.ctx.outputImport $v.Import>"
 :	:	<@end>\
 :	<@end>\
 	contr "<@.ctx.import>"
